+++ deployments/local/_helpers/configuratorModificator.ts
