+++ test/e2e/_template-1/_helpers/configuratorModificator.ts
