+++ test/e2e/_template-1/dai/deploy.ts
@@ -1,7 +1,7 @@
 import { Deployed, DeploymentManager } from '../../../plugins/deployment_manager';
 import { DeploySpec, deployComet } from '../../../src/deploy';
-import { makeToken, makePriceFeed, makeFauceteer } from '../_helpers/deployment-token-infra';
-import { ConfiguratorModifierHelper } from '../_helpers/configurator-modificator';
+import { makeToken, makePriceFeed, makeFauceteer } from '../_helpers/deploymentTokenInfra';
+import { ConfiguratorModifierHelper } from '../_helpers/configuratorModificator';
 import { exp, wait } from '../../../src/deploy';
 
 export default async function deploy(deploymentManager: DeploymentManager, deploySpec: DeploySpec): Promise<Deployed> {
